Move mini editor component to components
