+++ lib/main/atom/components/miniEditor.ts
